+++ src/main/java/org/apichart/common/modules/context/CustomKeyStoreParam.java
@@ -1,4 +1,4 @@
-package org.apichat.common.modules.context;
+package org.apichart.common.modules.context;
 
 import de.schlichtherle.license.AbstractKeyStoreParam;
