+++ src/main/java/org/apichart/common/modules/context/LinuxServerInfos.java
@@ -1,14 +1,17 @@
-package org.apichat.common.modules.context;
+package org.apichart.common.modules.context;
 
+import org.apache.commons.lang3.StringUtils;
+
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
 import java.net.InetAddress;
 import java.util.List;
-import java.util.Scanner;
 import java.util.stream.Collectors;
 
 /**
- * 用于获取客户Windows服务器的基本信息
+ * 用于获取客户Linux服务器的基本信息
  */
-public class WindowsServerInfos extends AbstractServerInfos{
+public class LinuxServerInfos extends AbstractServerInfos{
 
     @Override
     protected List<String> getIpAddress() throws Exception {
@@ -44,20 +47,19 @@ public class WindowsServerInfos extends AbstractServerInfos{
         //序列号
         String serialNumber = "";
 
-        //使用WMIC获取CPU序列号
-        Process process = Runtime.getRuntime().exec("wmic cpu get processorid");
+        //使用dmidecode命令获取CPU序列号
+        String[] shell = {"/bin/bash","-c","dmidecode -t processor | grep 'ID' | awk -F ':' '{print $2}' | head -n 1"};
+        Process process = Runtime.getRuntime().exec(shell);
         process.getOutputStream().close();
-        Scanner scanner = new Scanner(process.getInputStream());
 
-        if(scanner != null && scanner.hasNext()){
-            scanner.next();
+        BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
+
+        String line = reader.readLine().trim();
+        if(StringUtils.isNotBlank(line)){
+            serialNumber = line;
         }
 
-        if(scanner.hasNext()){
-            serialNumber = scanner.next().trim();
-        }
-
-        scanner.close();
+        reader.close();
         return serialNumber;
     }
 
@@ -66,20 +68,19 @@ public class WindowsServerInfos extends AbstractServerInfos{
         //序列号
         String serialNumber = "";
 
-        //使用WMIC获取主板序列号
-        Process process = Runtime.getRuntime().exec("wmic baseboard get serialnumber");
+        //使用dmidecode命令获取主板序列号
+        String[] shell = {"/bin/bash","-c","dmidecode | grep 'Serial Number' | awk -F ':' '{print $2}' | head -n 1"};
+        Process process = Runtime.getRuntime().exec(shell);
         process.getOutputStream().close();
-        Scanner scanner = new Scanner(process.getInputStream());
 
-        if(scanner != null && scanner.hasNext()){
-            scanner.next();
+        BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
+
+        String line = reader.readLine().trim();
+        if(StringUtils.isNotBlank(line)){
+            serialNumber = line;
         }
 
-        if(scanner.hasNext()){
-            serialNumber = scanner.next().trim();
-        }
-
-        scanner.close();
+        reader.close();
         return serialNumber;
     }
 }
