+++ src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.SpringApplicationRunListener=org.apichat.common.AbstractApplicationRunListener
+org.springframework.boot.SpringApplicationRunListener=org.apichart.common.AbstractApplicationRunListener
